Rollup DB changes to 1280 for initial database.
This updates the InitializeMythSchema() function to create a new
database at DB schema version 1280 (current schema version) so new users
will not be asked to allow a DB upgrade.  This is mainly a "dry run" of
a DB schema rollup, and I will do another just before 0.25 release.

Note that there are quite a few more tables than there were in the 1226
schema due, primarily, to the inclusion of MythVideo in core MythTV
(and, therefore, the inclusion of all of MythVideo's tables).  This also
means there's a lot more initial data.

This rollup also changes the schema initialization to specify the
ENGINE, since our code is primarily written with MyISAM in mind and
MySQL 5.5 changes the default ENGINE used when none is specified by the
CREATE TABLE command to InnoDB.  When we have come up with a plan for
transitioning to InnoDB and have removed as many MyISAM'isms as
possible, we will include a DB upgrade that converts existing databases.
Including the ENGINE in CREATE TABLE statements serves to ensure more
consistency among DB schemas.  An upcoming patch will change the
existing DB upgrades to specify ENGINE so that users who upgrade to
MySQL 5.5 are less likely to have mixed-ENGINE schemas where old tables
use MyISAM (as assumed by MythTV) but new tables use InnoDB.

A blank DB schema created with the previous 1226 rollup and upgraded to
1280 compared to the 1280 rollup was identical, save the DB-backup
information (since no DB backup occurred with the 1280 rollup).