Use lookup table for comparison operators.
This is a bit of an optimization of 7768ebc55dd, where the decision on
how to compare two different classes is now only performed once, at
which point the comparison is written into a lambda function, and stored
to a lookup table.  Further comparison of that type will use the
comparison method in the lookup table directly, rather than work through
the full decision tree each time.