Command line parser cleanup
Alters the 'long help' output to use the longest option string for a
given CommandLineArg when displaying relationships.

Adds proper reference counting of CommandLineArg and a destructor to
MythCommandLineParser to ensure instances are all freed properly in
the event the parser is deleted by the calling application.

This bumps the ABI.