Don't test linkability for the QtScript check.
Since we don't have a proper qmake-generated Makefile with proper Qt
environment set up, trying to do a compile-and-link test in configure
on some platforms seems to be far more trouble than it's worth, so just
do a quick compile test.  I have no idea why the QtDBus compile-and-link
test actually works, but I'm guessing it's because the QtDBus package is
closer to core.