Simplify the Qt configure checks.
Simplify the Qt configure checks a bit and make them more consistent.

This also changes the QtWebKit check to use pkg-config --exists rather
than --atleast-version 4.6.0 .  Since we now require Qt 4.6.0+, if
QtWebKit exists, it's always going to be at least version 4.6.0.

And, this changes the QtDBus check to be a non-linking check.  Since
we're not checking compile or linking when pkg-config says these Qt
modules exist, there's no real need to do the linking test in our
not-properly-configured-by-qmake limited environment when pkg-config
isn't available.  It may be that the failing Windows link test was
disabling qtdbus on Windows (which is fine, since qtdbus support is
optional--and Qt doesn't support DBus on Windows), but if we were
relying on this invalid test we may need to actually fix the code to do
the right thing in Windows (where that may mean we use the QtDBus stubs
I assume Qt installs on Windows and fall back appropriately or protect
the DBus code with some preprocessor tests).