Remove use of signals and slots from protocol server.
The rewritten delete thread in libmythprotoserver was intended to be
ignorant of any metadata pertaining to the file it was deleting, and
simply implement the slow delete and follow symlinks features.  Signals
would be fired off to indicate success or failure of deletion, so other
code could handle removal of the metadata from the database.

This reworks that behavior, replacing the deletestruct with a
DeleteHandler class, and replacing the signals with DeleteSucceeded()
and DeleteFailed() virtual methods.  In order to handle metadata,
subclass this and overwrite those two methods.  FileServerHandler
accepts it using the HandleDeleteFile() method.