Merge the former MythVideo translation files into core.
This merge the MythVideo translation files into the Mmythfrontend translation files now
that MythVideo has been merged into core.

This is the last one of the planned commits related to this. Any additionnal commit, if needed,
will be done on a per language basis.

Thank you to Kenni Lund and Nick Morrott for making sure this went smothly.