Updated MythFrontend Spanish translation from Juan Carlos Romero.
There was some problems with the plural translation of some strings which I have tried
to fix as best as I could but they will have to be revised (the plural had been put in
parenthesis just like the original source text instead of putting it in the plural position).

I also entered or fixed some strings which have to be copied verbatim for all translations but
had not been by a previous contributor.

I'll send you and the rest of the Spanish translation team the list of strings which should be
rechecked for correctness.

Thank you Juan Carlos!