Support Internal Firewire channel change with analog capture.
Allows one to specify "Internal" (minus quotes) as the channel change command.  Will then use MythTV's internal firewire device control to change the channels.

Adds two new DB fields in cardinput, one for changer device (The GUID for the box, exactly as specified when one sets up a firewire tuner (no 0x)) and the changer model, also exactly as specified for a firewire capture device (I use "MOTO GENERIC" without quotes).

I am not committing a UI for the GUID and model portions of this.  I have one, but figuring out how to get the dropdown for GUID and Model to appear only when an internal channel change is configured has so far escaped me.

I've been using this for a few weeks with perfect function.  This should be a nice way for new users of HD-PVRs, ivtv, framegrabbers, etc. to configure channel change without having to bother with a script.