Add nice/ioprio to MythSystem
Allows a MythSystem user to specify nice and io priority values that a
task should run at.  These will be set after the fork, so they will not
disrupt instances of a shared thread that would otherwise not allow the
current mechanism of niceing the thread prior to running MythSystem.