MythMusic: Update the track length to match the actual play time.
TabLib has problems with calculating the track lengths for VBR tracks if the
track doesn't contain a Xing header so the displayed track length can be wildly
wrong. This just checks the actual time a track played against what we have
in the DB and updates the DB if it is wrong.