Services APIs: Return Source ID when creating a source.
Now the APIs are fairly consistent and return relevant information when attempting to create a working capture setup.  It works more or less like this:

Call the /Capture/AddCaptureCard method with the appropriate arguments.  The Card ID for the created card is returned.
(Optionally: Call the /Capture/GetDDLineups method to pull in the lineups configured on Schedules Direct and get the lineupid for the next step)
Call the /Channel/AddVideoSource method with the appropriate arguments.  The Source ID for the created source is returned.
Call the /Capture/AddCardInput method with the appropriate arguments, the most important of which are the return values of the last two/three steps.  The Input ID of the created input is returned.

If the right info has been fed to the backend, upon restart you have a working capture setup.

Lots and lots remaining to be done-- query the backend for available capture card info, support XMLTV, channel scan, support channel fetching, etc.  But a halfway decent start that at least allows me to continue tinkering with a personal project for easy mac backend setup.