Fix spamming non-pipe select
It is necessary to reset the contents of rfds and efds both every time before
running select.  As this was missing from the loop, unexpected breakage
occurred.  Also, don't say we've received data if the rval <= 0 (i.e. no
results or an error).

This is inspired by the patch in #7844.  Closes #7844.