Adds the ability to remap action bindings in button lists with patch from Ronald Frazier.
While perhaps universally useful the real reason for this enhancement is to address the loss of arrow acceleration when we allowed themes to use horizontal layouts. To this end I've attempted to add the mappings where needed in MythCenter-wide. Please open tickets for any places I missed so they can be fixed.