MythDialogBox: Allow an hierarchical menu to be passed to the menu dialog.
This updates the menu popup dialog to accept a menu tree which the dialog
will allow the user to automatically move up and down the menu tree when
left/up and right/down are pressed. The OSD menu already allows this but in a
much more complicated way. Future commits will update some of the menus that
have submenus to use the new features.