Fix DTS playback under some circumstances.
When playing a DTS-HD track on a system only supporting DTS core (like when using SPDIF) we would try to decode the DTS-HD track which wouldn't work as FFmpeg doesn't support it. Instead on those system we bitstream the DTS core found in those DTS-HD stream.
When selecting which audio track to play, we don't even try to search for DTS-HD unless the hardware supports it.
Audio framework when being fed DTS (HD or not) will only use DTS core if it finds that the hardware doesn't support it. For the audio framework to attempt bitstreaming DTS-HD requires the hardware to support it