freemheg: Support bitmap backgrounds used by BBC/Freesat
Add support for bitmaps with a content hook of type 5 which are
transmitted by the BBC on FreeSat & FreeView.

Type 5 bitmaps are identical to type 2, an MPEG I-frame, but are intended
to be used as a background onto which the MHEG content and video are rendered.

At present when using the BBC 'Red button' service the background is all
black, missing the background image and red/green/yellow key legend
present on all other STB implementations.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>