MythUI: Use QTextLayout in mythuitext.
The primary purpose of this change, is to leverage QTextLayout's internal
tracking of the cursor position for use by mythuitextedit.  The cursor
position should now be accurate even when using non-mono-spaced fonts.
Multi-line textedits are now rendered correctly.  If the text being
edited does not fit in the allowed area, it will be automatically scrolled
as necessary, to keep the cursor in view.

QTextLayout supports a FormatRange of QTextCharFormat which would
allow us to have "spans" of differently formatted text within a textarea.
This patch does not add that support, I am just pointing out the possibility.