Make metadata grabber paths relative to PREFIX
The selected metadata grabber scripts are stored in system-wide settings
as an absolute path to the script.  This causes problems on
heterogeneous systems that may have multiple installs using different
shared folders.  This changes the path to be relative to the shared
folder, including a schema bump to modify existing settings.  This will
break on any system where the grabber has been manually altered to a
script outside the expected path.