Allow creating backend connections without the event socket
Add an optional parameter to MythCoreContext::ConnectToMasterServer()
to allow disabling the event socket connection.  Normally, we want
to open the event socket, but the forthcoming mythutil app just
needs to open a quick connection, send a message, read the response,
and exit.

This commit also adds a 'MYTH_APPNAME_MYTHUTIL' define for mythutil to
prevent a full recompile once it is committed.

This changes MYTH_BINARY_VERSION so make clean, etc..