HDHR Prime: Allow Prime tuners with CableCard to fetch channels.
With this commit, HDHomeRun Prime tuners are fully configurable
with no extra hoops to jump through.  If a a channel fetch is
requested, and if a HDHomeRun tuner is a member of the source,
the device will be checked for the presence of an OOB tuner. If
one is present, OOB tuner lock is checked.  If both are true,
channels are fetched from the listings source.  This has the
advantage of also ensuring that if a Prime tuner is used *without*
a functioning cablecard, it will behave as a traditional HDHR.

This addition is a minor hack, as it involves contacting the HDHR
rather than just checking the card type, but it will definitely
serve until the OCUR tuner type is ready, and then we can encourage
users to move to that tuner type to use the DRI interface.

This concludes the patches necessary to get the HDHR Prime
configuration experience where it should be for a release.  One
can now create all three tuners without issue, and fetch channels
without the use of a dummy tuner or other tricks.