Fix MythPlayer::GetRawVideoFrame() for mythcommflag.
Fix MythPlayer::GetRawVideoFrame() to return the correct frame number.
This change fixes mythcommflag so that it will return consistent results
across multiple runs.  It should also improve commercial detection
accuracy, as the frame numbers identified by the different methods will
actually be precise--allowing for better comparing the results of the
different methods.

Although this commit has my name for the signed-off line, I didn't do
much.  Thanks to Kevin Ross for the debugging and the patch, to Taylor
Ralph and Mark Kendall for actually reviewing the patch, and to all the
users and devs who ran with the patch to test it (and saw
greatly-improved commercial flagging results).

Signed-off-by: Michael T. Dean <mdean@mythtv.org>