Fix MythCommandLineParser::SetValue
Previously, this would store the new value to the CommandLineArg
instance, but would not mark the instance as having been used.  Thus, if
the argument had not actually been used on the command line, attempting
to retrieve the value would result in getting the default.  This
requires a version bump, so it also carries staging for a SetDeprecated
method to CommandLineArg.

Fixes #10083.