Fix dblogging deadlock during shutdown
Refs #10032.

If the program starts and exits while we are waiting for 100ms for the db to
be ready, we got into a deadlock due to a logging message being expedited while
the db queue is locked.  Message expediting only happens during shutdown.  The
way around this is to briefly unlock the queue while checking the database to
remove the possibility of deadlock.

An alternative fix would be to fix the delays to no longer use the conditional
waits, but that goes against the redesign goals that Daniel had earlier.