MythOpenGLPainter:  Render using the same engine for both Rect and RoundRect.
Even when RectanglesAreAccelerated() returned false, we where still using
opengl to render Rect, but not for RoundRect.  This resulted in
inconsistencies.

The correct solution is probably to figure out why the two different
rendering algorithms (OpenGL vs. Qt) produced different results, but that
was too much to get into right now.