Don't manipulate ProgList MythUI elements outside the UI thread
FillItemList is called both in the UI thread and outside the
UI thread.  Outside the UI thread, 'allowDisp' is false.  We don't
need to call m_progList->Reset() inside FillItemList because it
is called inside UpdateDisplay() which is called from FillItemList
when updateDisp is true.

This commit also refactors the two UpdateDisplay() methods a little
and adds a new RestoreSelection() method to perform the position
restoration that was originally in the 2nd UpdateDisplay() method.

Fixes #10073.