Unlock the db queue mutex during InitCon in the startup
If the program is exiting quickly (before the database is ready), the db queue
is still locked, and then the messages bypass the queue.  When it goes to
initialize the database connection, the code in mythdbcon logs a message, which
then tries to acquire the lock again, causing a deadlock.

To get around that, unlock the mutex around the call to InitCon().  I think it
would be better to re-re-think the waiting code so it's not using that mutex
so the mutex can get locked only after the connection is ready, removing this
and the previous unlock/relock.

Refs #10032.  Hopefully this will fix the deadlocks.