Stubs out SCTE tables.
SCTE shares a lot of tables with ATSC, but there are a few tables in SCTE streams that differ from ATSC.

When we implemented ATSC support in MythTV SCTE tables weren't very interesting because they were carried
out of band and none of the consumer cards could decode them. Now Comcast is placing these in the QAM stream
and the tables may also be available from OCUR tuners, so they are a bit more interesting.