Add partial SpliceInformationTable support.
A long time ago we started seeing CUEI descriptors and we categorized
those a being indicitive of an SCTE 65 stream. It is a table specified by
SCTE, but not as part of the US Cable PSIP standard (SCTE 65). It is
actually part of SCTE 35 which is used globally to indicate program
splice points.