DVB Radio: Refactor generation of dummy video frames.
- Fill the video buffers completely on each call instead of creating one
frame for each call to AvFormatDecoder::GetFrame. Audio or MHEG only
streams may have very low bitrates and the player will otherwise be
starved of buffers.

- Don't create a dummy frame, just ask the player to clear the returned
frame. This will allow the VideoOutput method to clear frames in GPU
memory OR just ignore the dummy frame when it is not required for
display (e.g. VDPAU, VAAPI, OpenGL, DXVA2).

Ref #9824