Add a new EventCmdScreenType MythSystemEvent.
Add a new System Event that is sent out whenever a new screen is
created or an existing screen is destroyed.  This will allow users
to take external actions whenever an application enters or leaves
a screen/dialog.

The system event is SCREEN_TYPE and the setting name is
EventCmdScreenType.  The first argument in the %ARGS% value is
either 'CREATED' or 'DESTROYED' depending on whether a MythScreenType
was created or destroyed.

Fixes #9395.