Subtitles: Add initial forced subtitle support.
This enables forced subtitles when they are part of the current,
automatically selected subtitle stream (i.e. if your language is
English, an English subtitle track is automatically selected in the
existing code. Any forced subtitle detected will now be displayed).

I will add an OSD menu option to toggle forced subtitles.

This will not work for blu ray material where the forced subtitles are
on a different track entirely.

There are still some intermittent issues with DVD material that seem to
be related to the existing code. In these cases the forced subtitles may
no display.