Fix mythtranscode filter use.
The setting for mythranscode filters has existed for a long time,
but it doesn't look like it was ever fully hooked up or else it
was removed somewhere along the line during a code restructure.

This adds a call to videoFilters->ProcessFrame() in
MythPlayer::TranscodeGetNextFrame() to process the last decoded frame.

Fixes #8533.