Fix a crash caused by background animated image loading.
The background animated image loader was calling SetImages() which
triggers non thread-safe routines in MythUI.  Instead of calling
SetImages() in the background thread, return the image QVector
in the ImageLoadEvent and set the images and delays in the UI
thread similar to how we set non-animated images.

Fixes #9388.

NOTE: This changes the binary API, so make clean, etc., whatever
you normally do.