Fix race condition and potential segfault in MythDownloadManager.
Put a lock around the creation of the MythDownloadManager to make
sure we don't get two created by a potential race condition.

Put a lock around access to the QNetworkDiskCache when called
via MythDownloadManager::GetLastModified().

Fixes #9231 using ideas from Jonatan from comhem.se, but most of
the patch (and potential issues) are mine.