Attempt to reduce stupid warnings on -ppc buildslave
The linux-ppc buildslave is using a newer version of gcc than the others,
so it has pointed out some extra warnings.  The warning that this should
permanently remove is:

./../external/FFmpeg/libavcodec/get_bits.h:305:45: warning: variable ‘re_cache’ set but not used [-Wunused-but-set-variable]

This is caused when using the skip_bits() function as the OPEN_READER() macro
creates a re_cache variable, but nothing uses it in that specific call flow.
The fix is to specifically reference it in the CLOSE_READER() macro so it is
always used.  A void cast reference should be enough to make gcc happy, but
yet not affect the output code.

This has been tested on the older build systems, but not yet on the ppc slave.
If it goes awry, I'll revert it.  We should look at pushing this upstream (to
one or both upstreams) if it proves to work well.