Remove many superfluous warnings from goom code
These warnings started popping up as soon as we actually set the CFLAGS
correctly for compiling the C parts with warnings on and optimization, etc.
As this code is imported, this can be considered to be rude hacks to shut
gcc up, as we don't want to take the effort to clean up the code in reality.