Fix potential segfault in MythUI teardown.
When the main screen stack is deleted, NULL out the pointer
tracking it in MythMainWindowPrivate.

In MythScreenStack::PopScreen(), check that the pointer
returned by MythMainWindow::GetMainStack() is non-NULL before trying
to use it to check the top screen on the main stack.

I believe this cures the segfault reported in IRC by several people
that occurred in QRect::width() when called by MythUIType::SetRedraw()
from MythScreenStack::PopScreen().  It was hard to get decent debugging
output on this, but it appears we were trying to access the already-
deleted main screen stack when deleting the background screen stack.