Add a video performance test to mythavtest.
The test disables audio, subtitling, OSD, video filtering (including
deinterlacing) and simply tries to decode/display frames as fast as
possible. Should prove useful for comparative and regression testing. I
may add an extra option to enable deinterlacing to assess its impact on
performance. The playback profile (decoder, video output) will be driven
by the mythfrontend settings.

Specify the file to test with -f or --file (any directly accessible file
or valid url) and request a test with -t or --test.

Optionally adjust the test duration with -s or --seconds (defaults to 5)
or request a decoder only test with -d or --decodeonly (frames are just
decoded and not otherwise processed or displayed).

Look to the log output for the results.

N.B. many display methods will be rate limited due to sync-to-vblank
being enabled. This is unavoidable in some cases (e.g. VDPAU), can be
turned off for others (e.g. most linux OpenGL implementations) and
others will need changes to the code to allow disabling them (e.g.
XVideo, OpenGL on OSX).