OpenGL Video: Add a new 'opengl-lite' video display profile.
This is a slightly customised OpenGL render profile that will prefer the
use of certain OpenGL extensions (GL_MESA_ycbcr_texture and
GL_APPLE_ycbcr_422).

These extensions are used to upload video textures to GPU memory. They
are usually faster when available but don't allow full colourspace
control or the use of OpenGL based deinterlacers.

In short, the profile is more efficient on certain chipsets but at the
cost of full functionality.

It should prove useful for OpenGL output on Intel and ATI GPUs and OSX.
There is no support with Nvidia drivers.