MythUI: Pay attention to alignment in <minsize> areas.
The main benifit of this is seen when <minsize> is used outside of
buttonlists.

For example, users can select among several different date/time formats
which range from being pretty short, to fairly long.  If the themer wanted a
background shape to show under the date/time, they had to allow for the
longest of date/time formats which did not always look good, if a short
format was selected.  Using <minsize> a themer can now have the background
shape auto-sized to the text length, while still throwing it all into a
corner:

    <group name="dateGroup">
	<area>100%-500, 20, -20, 40</area>
	<minsize>20, 100%</minsize>

	<shape name="dateshape">
	    <area>0, 0, 100%, 100%</area>
	    <minsize>20, 100%</minsize>
	    <type>roundbox</type>
	    <cornerradius>15</cornerradius>
	    <fill color="#000000" alpha="255" />
	</shape>

	<clock name="datetime">
	    <area>10, 10, -10, -10</area>
	    <minsize>20, 100%</minsize>
	    <font>basesmalllightgrey</font>
	    <format>%DATE%   %TIME%</format>
	    <align>right</align>
	</clock>
    </group>