Allow bad Metadata tags to be bypassed
This is in response to a change in the TMDb API that resulted in the
grabber returning bad data for the 'lastupdated' field.  If a field
returns bad data, skip over it and continue processing the rest of the
result.

(cherry picked from commit 15b0ffc383b87173708ccd867e0f0fdddceec4da)