previewgen: Fix handling of --infile
Currently if the infile name looks like a myth recording title with a valid
channel ID and start time then the file is only searched for in the
recordings folder and so is consequently not found.

This fix checks if the file exists and if so doesn't attempt to extract
the channel ID and start time with the result that the file is not searched for in the recordings folder.

Signed-off-by: Lawrence Rust <lvr@softsystem.co.uk>
Signed-off-by: Stuart Morgan <smorgan@mythtv.org>