Fix waiting for the logging table on upgrade.
A logic-o/typo in [aab13b393] meant that we stopped waiting for the
logging table as soon as we had a MythCoreContext, which resulted in a
spew of log messages saying that the logging table didn't exist for
users who were upgrading their databases from versions before the new
logging API or creating new databases.

This fixes the wait code so that users won't be confused by/blame failed
upgrades on the (harmless) messages that simply meant we were unable to
log to the database, yet.