Don't prepare DB logging query if aborted.
If aborted is set, we're shutting down, so don't prepare the database
logging query.  This will prevent errors, for example, if the user
cancels a DB upgrade.  And, since the while loop will prevent actually
writing any log messages on aborted, there's no reason to do any of the
preparation code.

This looks like a big change due to whitespace changes but is really
just inserting an if (!aborted) around the main block of code in
DBLoggerThread::run().