ProgramInfo: get the video, audio, and sub types from program.
This constructor is only used in one spot, when calculating the guide.
For us to support more advanced features in our own guide, and third
party guides using the Services API, these properties need to be present.

In a MythUI guide, this will allow for statetypes for HD, stereo, dolby, etc.

Services API addition to follow.