Services API: Add Video, Audio, and Subtitle props to Program contract.
This allows an API consumer to pull up the guide and get the bitmasks
for video, audio, and subtitle properties.

I went back and forth on whether to consider these attributes part of
details of the base view, but settled on them being a default.  If one is using
the GetProgramGuide API, one needs these attributes to build the guide visually.
One can then get the further details with the API to get full details on a program
if the item is clicked/selected/tapped/whatever.