MythUIButtonList: move arrow states to the top.
The selected button is drawn on top of any non-selected buttons, but it was
also drawn on top of any arrows.  In the vast majority of cases, the
arrows should be on top.

It would be nice if the selected button was automatically stacked on top of
the other buttons, and yet still leave it up to theme at which layer the arrows
should be drawn, but that is a far more invasive change.