Use ProgramID for Video Files.
More accurately, extend the programinfo constructor for videos
to add programid. Load Video ID into programID when creating
a ProgramInfo for the player.  This will allow the frontend status
query to display the video ID, which can then be used as a basis for
a metadata query of the backend Video Service API.

Bumps the library API, make distclean.