VideoList/UPNPScanner: Add a dynamic sub folder type.
Add a dynamic subfolder type that is updated when selected. This is used
by the UPNPScanner code and completes the conversion from a complete
UPnP scan to scanning on demand.

The overall approach is convoluted as we move from MediaServerItem's in
the scanner, to video metadata and ultimately to a MythGenericTree (and
the button list...). The actual interface is still clunky as the item
needs to be selected to trigger the scan and then selected again to move
into the new item.