Playback: Fix DVD and Blu Ray seeking.
The main problem here was that when paused, the seek position would
fallback to the current speed (i.e. 0) and playback would resume from
the beginning of the title.

With DVD playback, we were also asking the dvdnav object to seek to the
nearest cell - which led to some sizeable search discrepancies depending
where playback was relative to the nearest cell. We now ask for a more
accurate search.

N.B. I also tried to use a relative time search to the nearest cell
followed by a smaller absolute time search (as outline in the dvdnav
api) but this seemed to be even less accurate.

Refs #9305