Fix DVD and Blu Ray seeking.
The main problem here was that when paused, the seek position would
fallback to the current speed (i.e. 0) and playback would resume from
the beginning of the title.

With DVD playback, we were also asking the dvdnav object to seek to the
nearest cell - which led to some sizeable search discrepancies depending
where playback was relative to the nearest cell. We now ask for a more
accurate search.

Backported from master 4b0803be0c0974c184

Fixes #9305