Content Service: Generate screenshots at the requested size.
This is an artifact from way back in 2007/8 when lots of MythXML
preview generation calls were causing BE deadlocks.  Since we believe
we have vanquished the deadlocks and because this API is now
beuing used by third party applications that need high quality imagery,
generate the images at the requested dimensions.

Has the side effect of making the item detail screen in mythweb
look WAY better (not to mention the backend web setup landing page,
and my own third party apps).