Services API: Add initial Status service to MythFrontend.
- removes the 'old' status method from MythFEXML.

- adds a new frontend service to fetch the status which should respond
to requests for xml, json etc

- the previous approach of having the UI send the status back via an
event would not work with the services structure (the objects created to
respond to requests do not have their own thread and are transient), so
I instead added a state tracker singleton to libmythui that is refreshed
on request. This could obviously be extended for other purposes.

- more MythFEXML methods will be moved over in due course and the whole
frontend remote interface reviewed (we now have a frontend service, the
http status page and the other MythFEXML methods).