libmythmetadata: fix rare occasions when artwork was cleared out.
In "absolute last resort fallback mode," when trying to find at least
series level artwork for a generic TV episode, we were accidentally passing
the lookup object for the title search back as a result-- this lookup
only contains partial info, and not full artwork, and when there was
existing, full artwork for season 1, the incomplete artwork for the
"0" season was being overwritten in its stead-- at most, this was a
banner as this lookup was only meant to be for producing a search
selection list.

Fix the issue by taking the single series result and performing
one more lookup on it as a faux "1x01," which will produce correct,
full artwork.

Confused yet?