Add --jumppoint and --runplugin to mythfrontend
This adds two new options for mythfrontend.  --jumppoint will
immediately jump to the specified point when the frontend has finished
loading.  --runplugin will jump to the base window of the specified
plugin.  Both will terminate if an invalid name is given.

This removes the old behavior of allowing plugin access through
alternate executable file names, or extra arguments on the command line.

Closes #10135

Signed-off-by: Raymond Wagner <rwagner@mythtv.org>