Allow --jumppoint and --runplugin to print available destinations
This adds methods to MythMainWindow and MythPluginManager to allow
enumeration of available jump points and plugins, and uses it to list
them when an invalid one is supplied on the command line. This causes a
ABI bump.

Also, this moves the code to load a specific plugin to just prior to
normal program init in the frontend, ensuring certain background tasks
get initialized.

Refs #10135