Extend the Frontend Status.
- adds "currentlocation" to the main UI response.
- adds "audiosyncoffset" during playback.
- adds a list of chapter times (in seconds) if available during
playback.

The chapter times are stored as a QList within the overall status map
but, to avoid potentially invalid formatting changes to the services API
serialisers, they are then copied across to a different property within
the actual frontend status class. Converting the json serialiser to cope
with nested variant lists is trivial but the same change to the xml
serialiser would potentially give invalid xml.

N.B. there is no useful list for either titles or angles, though a title
list could replicate the bluray OSD entry that uses the title length as
a description.