Playback: Improved triggers for low bitrate optimisations.
The optimisations added in d207a0bddddac are being triggered
unnecessarily. This is due to a combination of different initial
thresholds in that commit and lowering the acceptable bitrate in
a1022c6f508c3.

The fix is twofold:-

- adjust the initial bitrate thresholds to ensure any optimisations are
only triggerred for streams with a bitrate below 500Kb/s.
- any video codec scanned in AvFormatDecoder that has a zero bitrate is
forced to 500Kb/s. This is an extension of the previous hack for H.264
streams.

Overall, this ensures that the previous behaviour (initial fill size of
32KB, rising as needed) is preserved for anything with a video stream
and generally in the vast majority of cases.

Fixes recent playback of avi files and, amongst others, BB HD (both
detected as zero bitrate).

Refs #9824.